introduce module for reporting util functions
